create cliente

+++ database/migrations/2021_01_07_164008_create_clients_table.php
@@ -23,7 +23,7 @@ class CreateClientsTable extends Migration
             $table->string('numar');
             $table->integer('codPostal');
             $table->string('persoanaContact');
-            $table->integer('numarTelefon');
+            $table->bigInteger('numarTelefon');
 
             $table->timestamps();
         });
